Add store and storeSubscription to defaultContextTypes.
This change allows the default store and subscription passed in the
react-redux Provider context to pass through to ShadomDOM children.  The
Provider does permit the default store and subscription to be changed.
This patch does not support non-standard store names.  If necessary, a
user can generate a new, custom ShadowDOM object using `withContext` to
support non-standard store names.